Add done button and handle action
